+++ cmd/deploy/error.go
@@ -6,6 +6,10 @@ var invalidConfigError = &microerror.Error{
 	Kind: "invalidConfigError",
 }
 
+var invalidFlagError = &microerror.Error{
+	Kind: "invalidFlagError",
+}
+
 var invalidArgError = &microerror.Error{
 	Kind: "invalidArgError",
 }