+++ cmd/runner.go
@@ -96,12 +96,8 @@ func (r *runner) persistentPreRun(ctx context.Context, cmd *cobra.Command, args 
 
 	latestVersion, err := updaterService.GetLatest()
 	if updater.IsHasNewVersion(err) {
-		fmt.Fprintf(r.stderr, "If you know what you are doing you can disable this check by exporting %s=%s\n", env.DevctlUnsafeForceVersion.Key(), project.Version())
-		fmt.Fprintf(r.stderr, "Current version:  %s\n", project.Version())
-		fmt.Fprintf(r.stderr, "Latest version:   %s\n", latestVersion)
-		fmt.Fprintf(r.stderr, "Please update your %s with \"%s version update\"\n", project.Name(), project.Name())
-
-		return microerror.Mask(err)
+		return fmt.Errorf("version %[2]s of %[1]s is released; this is %[3]s: update with `%[1]s version update`, or run this version anyway with %[4]s=%[3]s",
+			project.Name(), latestVersion, project.Version(), env.DevctlUnsafeForceVersion.Key())
 	} else if err != nil {
 		return microerror.Mask(err)
 	}